+++ src/main/resources/db_construction/tables/clinics.sql
@@ -1,12 +1,12 @@
 create table mole.clinics (
-  clinic_id INTEGER GENERATED ALWAYS AS IDENTITY (START WITH 1) PRIMARY KEY
+  clinic_id BIGINT GENERATED ALWAYS AS IDENTITY (START WITH 1) PRIMARY KEY
   , clinic_date DATE
-  , clinic_location_id INTEGER
+  , clinic_location_id BIGINT
 
   , created TIMESTAMP DEFAULT CURRENT_TIMESTAMP NOT NULL
-  , created_by INTEGER NOT NULL
+  , created_by BIGINT NOT NULL
   , modified TIMESTAMP DEFAULT CURRENT_TIMESTAMP NOT NULL
-  , modified_by INTEGER NOT NULL
+  , modified_by BIGINT NOT NULL
   , deleted BOOLEAN DEFAULT FALSE NOT NULL
 )
 ;
