+++ src/main/resources/db_construction/tables/clinicslots.sql
@@ -1,15 +1,15 @@
 create table mole.clinic_slots (
-  clinic_slot_id INTEGER GENERATED ALWAYS AS IDENTITY (START WITH 1) PRIMARY KEY
-  , clinic_id INTEGER
+  clinic_slot_id BIGINT GENERATED ALWAYS AS IDENTITY (START WITH 1) PRIMARY KEY
+  , clinic_id BIGINT
   , datetime TIMESTAMP
-  , agency_id INTEGER
-  , clinic_slot_outcome_id INTEGER
-  , claimant_id INTEGER
+  , agency_id BIGINT
+  , clinic_slot_outcome_id BIGINT
+  , claimant_id BIGINT
 
   , created TIMESTAMP DEFAULT CURRENT_TIMESTAMP NOT NULL
-  , created_by INTEGER NOT NULL
+  , created_by BIGINT NOT NULL
   , modified TIMESTAMP DEFAULT CURRENT_TIMESTAMP NOT NULL
-  , modified_by INTEGER NOT NULL
+  , modified_by BIGINT NOT NULL
   , deleted BOOLEAN DEFAULT FALSE NOT NULL
 )
 ;
