+++ src/main/resources/db_construction/tables/comments.sql
@@ -1,12 +1,12 @@
 create table mole.comments (
-  comment_id INTEGER GENERATED ALWAYS AS IDENTITY (START WITH 1) PRIMARY KEY
-  , case_id INTEGER
+  comment_id BIGINT GENERATED ALWAYS AS IDENTITY (START WITH 1) PRIMARY KEY
+  , case_id BIGINT
   , comment_text CLOB
 
   , created TIMESTAMP DEFAULT CURRENT_TIMESTAMP NOT NULL
-  , created_by INTEGER NOT NULL
+  , created_by BIGINT NOT NULL
   , modified TIMESTAMP DEFAULT CURRENT_TIMESTAMP NOT NULL
-  , modified_by INTEGER NOT NULL
+  , modified_by BIGINT NOT NULL
   , deleted BOOLEAN DEFAULT FALSE NOT NULL
 )
 ;
